src/requirements.txt
boto3==1.9.174
pyyaml==5.1.1
